Исправлены названия модулей, убран renderTemplate
