+++ src/mock/site-menu-items.js
@@ -1,4 +1,4 @@
-export const generateMenuItems = () => {
+export const generateSiteMenuItems = () => {
 
   return [
     {
