+++ src/view/waypoint-destination.js
