+++ src/view/waypoint-offers.js
